pyproject.toml
[project]
name = "book-api"
version = "0.1.0"
description = "API to handle and serve book data."
authors = [
    {name = "Juliano Choi",email = "julianochoi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlalchemy[aiosqlite] (>=2.0.40,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.5"
mypy = "^1.15.0"

[tool.ruff]
indent-width = 4
line-length = 120

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "T20",
  "FAST",
]

[tool.ruff.format]
indent-style = "tab"

[tool.mypy]
files = "app"
strict = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"